requirements.txt

Django==4.2
gunicorn==20.1.0
celery==5.2.7
djangorestframework==3.14.0
geventhttpclient==2.0.9

psycopg2-binary==2.9.5
redis==4.5.4
watchdog==3.0.0
requests==2.28.2
flower==1.2.0
websocket-client==1.5.1

# grpc
grpcio==1.54.0
grpcio-tools==1.54.0